+++ src/input.cpp
@@ -274,7 +274,7 @@ void dead_zone(r32* x, r32* y, r32 threshold)
 
 r32 dead_zone(r32 x)
 {
-	if (fabs(x) < UI_JOYSTICK_DEAD_ZONE)
+	if (fabsf(x) < UI_JOYSTICK_DEAD_ZONE)
 		return 0.0f;
 	else
 		return (x - UI_JOYSTICK_DEAD_ZONE) / (1.0f - UI_JOYSTICK_DEAD_ZONE);
